+++ posthog/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- feat: Stateless client can override event `timestamp` when capturing an event ([#297](https://github.com/PostHog/posthog-android/issues/297))
+
 ## 3.23.1 - 2025-09-30
 
 - fix: `PostHogStateless` now deduplicates `$feature_flag_called` events ([#293](https://github.com/PostHog/posthog-android/pull/293))